+++ src/api/resources.js
@@ -1,6 +1,4 @@
 import axios from 'axios';
-import FileSaver from 'file-saver';
-
 import { apiURLBase } from '../config';
 
 /**
@@ -32,14 +30,12 @@ export function getResourceDetail(resource, sourceTargetToken) {
   });
 }
 
-
 /**
  * Resource Download Endpoint
  **/
 export function getResourceDownload(resource, sourceTargetToken) {
   const resourceDownloadURL = resource.links.find(link => link.name === 'Download').link;
   return axios.get(resourceDownloadURL, { headers: { 'presqt-source-token': sourceTargetToken } });
-  // return axios.get('https://localhost/api_v1/targets/osf/resources/bad_id.zip/', { headers: { 'presqt-source-token': sourceTargetToken } });
 }
 
 export function resourceDownloadJob(downloadJobURL, sourceTargetToken) {
