+++ src/redux/saga/saga_root.js
@@ -1,7 +1,8 @@
 import {all} from 'redux-saga/effects';
 
 import {watchLoadTargets} from "./targets";
-import {watchSearch, watchSelectSourceResource, watchSwitchSource, watchSourceResourceDownload} from "./resources";
+import {watchSearch, watchSelectSourceResource, watchSwitchSource} from "./resources";
+import {watchSourceResourceDownload} from "./download";
 
 // notice how we now only export the rootSaga
 // single entry point to start all Sagas at once
